requirements.txt
dapr-agents==0.10.5
chainlit>=1.0.0
python-dotenv
websockets>=12.0
